refactor: put CSS into their own files and general re-organising
