+++ frontend/src/App.tsx
@@ -1,13 +1,12 @@
 import React from "react";
 import { Route, Switch } from "react-router";
 
-import { AuthProvider } from "./hoc/AuthContext/auth-context";
-import ProtectedRoute from "./hoc/ProtectedRoute/protected-route";
-import Layout from "./hoc/Layout/layout";
-
 import Homepage from "./components/Homepage/homepage";
 import Login from "./containers/Login/login";
 import Units from "./containers/Units/units";
+import { AuthProvider } from "./hoc/AuthContext/auth-context";
+import Layout from "./hoc/Layout/layout";
+import ProtectedRoute from "./hoc/ProtectedRoute/protected-route";
 
 const App: React.FC = () => {
   return (
