+++ frontend/src/components/Navigation/Header/header-styles.tsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const Header = styled.header`
+  height: 75px;
+  opacity: 0.75;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 99;
+  background-color: rgba(68, 89, 106, 0.8);
+  border-radius: 5px;
+  position: unset;
+  border: 2px solid yellow;
+`;
