+++ frontend/src/components/Navigation/NavigationItems/NavigationItem/navigation-item.tsx
@@ -1,19 +1,5 @@
 import * as React from "react";
-import { NavLink } from "react-router-dom";
-import styled from "styled-components";
-
-const List = styled.li``;
-
-const NavigationItem = styled(NavLink)`
-  display: block;
-  margin: 0 6rem;
-  color: rgb(254, 243, 115);
-  text-decoration: none;
-
-  &:hover, &:active {
-      color: white;
-  }
-`;
+import { NavigationItem } from "./navigation-item-styles";
 
 const navItem = ({
   link,
@@ -27,11 +13,20 @@ const navItem = ({
   clicked?: (event: React.MouseEvent<HTMLAnchorElement>) => Promise<void>;
 }) => {
   return (
-    <List>
-      <NavigationItem to={link} exact onClick={clicked ? async (e) => await clicked(e) : undefined}>
+    <ul>
+      <NavigationItem
+        to={link}
+        exact
+        onClick={
+          clicked
+            ? async (event: React.MouseEvent<HTMLAnchorElement>) =>
+                await clicked(event)
+            : undefined
+        }
+      >
         {children}
       </NavigationItem>
-    </List>
+    </ul>
   );
 };
