+++ frontend/src/components/Navigation/NavigationItems/navigation-item-styles.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const List = styled.ul`
+  list-style: none;
+  margin: 0 auto;
+  padding: 0;
+  display: flex;
+  align-items: center;
+`;
