+++ frontend/src/containers/Login/login-styles.tsx
@@ -46,7 +46,7 @@ export const InputButton = styled.input`
    font-weight: 300;
    color: #000000;
    text-shadow: 0 0.04em 0.04em rgba(0, 0, 0, 0.35);
-   background-color: #42cc8c;
+   background-color: yellow;
    text-align: center;
    transition: all 0.15s;
 
@@ -69,7 +69,7 @@ export const Button = styled.button`
    font-weight: 300;
    color: #000000;
    text-shadow: 0 0.04em 0.04em rgba(0, 0, 0, 0.35);
-   background-color: #00ffff;
+   background-color: chartreuse;
    text-align: center;
    transition: all 0.15s;
