+++ frontend/src/hoc/AuthContext/auth-context.tsx
@@ -1,6 +1,6 @@
-import React, { useState, useEffect, useContext } from "react";
-import { useHistory } from "react-router-dom";
 import axios from "axios";
+import React, { useContext, useEffect, useState } from "react";
+import { useHistory } from "react-router-dom";
 
 export interface AuthContextInterface {
   user: {
