+++ frontend/src/hoc/Layout/layout.tsx
@@ -1,17 +1,16 @@
 import * as React from "react";
 
-import Header from "../../components/Navigation/Header/header";
 import Footer from "../../components/Navigation/Footer/footer";
+import Header from "../../components/Navigation/Header/header";
 
-const layout = ({ children }: {children: React.ReactNode}) => {
-
-    return (
-        <React.Fragment>
-            <Header />
-            {children}
-            <Footer />
-        </React.Fragment>
-    )
-}
+const layout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <React.Fragment>
+      <Header />
+      {children}
+      <Footer />
+    </React.Fragment>
+  );
+};
 
 export default layout;
