+++ frontend/src/hoc/ProtectedRoute/protected-route.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
-import { Route, Redirect } from "react-router-dom";
+import { Redirect, Route } from "react-router-dom";
+
 import { useAuthContext } from "../AuthContext/auth-context";
 
 const ProtectedRoute = ({
