+++ package.json
@@ -7,7 +7,7 @@
     "start-data-import-service": "cd data-import && npm run build && npm run start",
     "postinstall": "run-s install:*",
     "start": "cd backend && npm run start",
-    "heroku-postbuild": "npm run install:backend && npm run install:data-import && npm run start-data-import-service"
+    "heroku-postbuild": "npm run install:backend"
   },
   "dependencies": {
     "commitizen": "^4.1.2",
